pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mopaint"
version = "0.1.5"
description = "Paint for marimo"
authors = [{name = "Vincent D. Warmerdam"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
    "anywidget>=0.9.2",
]
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering",
]

[project.urls]
Homepage = "https://github.com/koaning/mopaint"
Documentation = "https://koaning.github.io/mopaint/"
Repository = "https://github.com/koaning/mopaint"
Issues = "https://github.com/koaning/mopaint/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pillow>=9.0",
    "marimo",
    "mohtml",
]

[tool.hatch.build.targets.sdist]
include = [
    "/mopaint",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["mopaint"]

[tool.hatch.build.targets.wheel.sources]
"mopaint/static" = "mopaint/static"